pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "familator"
version = "0.1.0"
description = "NES/Famicom ROM Automated Text Extraction, Translation & Reinjection"
authors = [{name = "Matt-Retrogamer", email = "matt@retrogamer.dev"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Localization",
    "Topic :: System :: Emulators",
]
keywords = ["nes", "famicom", "rom", "translation", "retro", "gaming", "ai", "llm"]

dependencies = [
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
enhanced = [
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/Matt-Retrogamer/FamiLator"
Repository = "https://github.com/Matt-Retrogamer/FamiLator"
Documentation = "https://github.com/Matt-Retrogamer/FamiLator/blob/main/README.md"
"Bug Tracker" = "https://github.com/Matt-Retrogamer/FamiLator/issues"

[project.scripts]
familator = "scripts.run_pipeline:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["yaml.*", "requests.*"]
ignore_missing_imports = true
